updated HttpClientRepository - moved LifetimeTrackingProxiedHttpMessageHandler into Internal folder - added AssemblyInfo to expose internals to test project - updated repository to take HttpMessageHandlerFactory which is used as a fallback if no such name is registered, efforts are made to use HttpClientFactory before it comes to that
